Add a basic 'Add Brew' view
This adds both the GET and POST routing for this basic view. It actually
works, sticks a new brew into redis everywhere necessary, and then shows
it on the recent brews page. Great success!

Obviously error handling, where to redirect, etc. all need a lot of
cleanup but this is totally demoable at this point.