Get Socket.IO working smoothly
This is starting to really come together. We now successfully listen for
updateBrew and deleteBrew messages in the browser and respond
accordingly, either adding the brew to the page or deleting it if
necessary. Successfully tested this in Chrome and Firefox with a third
tab used for adding the latest brew.